pyproject.toml
[tool.poetry]
name = "autogit"
version = "0.1.0"
description = ""
authors = ["Chen Yang <yangchen@thinkmore>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.20"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
autogit = 'autogit.console:run'


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
